+++ tests/core/test_indicator_timeframe.py
@@ -17,7 +17,7 @@ class FakeIndicator(Indicator):
         return f"{self.indicator_name}_{self.period}"
 
     def _calculate_reading(self, index: int) -> float | dict | None:
-        return 100.0
+        return self.reading("open") + self.reading("close")
 
 
 def remove_indicators(candles: List[Candle]) -> List[Candle]:
@@ -45,3 +45,20 @@ def test_collapse_candles_minutes_t10(
     test = FakeIndicator(candles=minimal_candles, timeframe="t10")
 
     assert test.candles == minimal_candles_t10
+
+
+@pytest.mark.usefixtures("minimal_candles", "minimal_candles_t5")
+def test_collapse_candles_minutes_t5_partial(
+    minimal_candles: List[Candle], minimal_candles_t5: List[Candle]
+):
+    minimal_candles_t5[0].indicators = {"Fake_10_T5": 25631}
+    minimal_candles_t5[1].indicators = {"Fake_10_T5": 20069}
+    minimal_candles_t5[2].indicators = {"Fake_10_T5": 33201}
+    minimal_candles_t5[3].indicators = {"Fake_10_T5": 33184}
+
+    minimal_candles = remove_indicators(minimal_candles)
+    test = FakeIndicator(candles=minimal_candles[:3], timeframe="t5")
+    test.calculate()
+    test.append(minimal_candles[3:])
+
+    assert test.candles == minimal_candles_t5